requirements.txt
matplotlib==3.10.8
numpy==2.3.5
pandas==2.3.3
pillow==12.0.0
scikit-image==0.25.2
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
